extract domain
