+++ backend/Dockerfile.core
@@ -6,10 +6,12 @@ WORKDIR /app
 # Create necessary directories
 RUN mkdir -p /core
 RUN mkdir -p /sheepshead
+RUN mkdir -p /database
 
 # Copy only the necessary files for dependency resolution
 COPY ./core/go.mod ./core/go.sum ./core/
 COPY ./sheepshead/go.mod ./sheepshead/go.sum ./core/
+COPY ./database/go.mod ./database/go.sum ./core/
 WORKDIR /app/core
 RUN go mod download
 
@@ -17,6 +19,7 @@ RUN go mod download
 WORKDIR /app
 COPY ./core ./core
 COPY ./sheepshead ./sheepshead
+COPY ./database ./database
 
 # Build the application binary
 WORKDIR /app/core
