--- backend/core/domain/entity/README.md
@@ -1,1 +1,0 @@
-The `entity` folder contains structs that are persisted in a database.
