+++ backend/database/README.md
@@ -1,0 +1,17 @@
+# The Database Module
+
+This module provides the data models for the application.
+
+## Packages
+
+### Entity
+
+The `entity` package contains structs that are persisted in a database. These structs represent the core data models for the application.
+
+### Repository
+
+The `repository` package contains interfaces that define CRUD operations for entities. These interfaces abstract the data access layer, allowing for easier testing and flexibility in changing the underlying data storage implementation. Each repository interface typically includes methods for creating, reading, updating, and deleting records.
+
+### Implementation
+
+The `implementation` package contains implementations of the `repository` package interfaces. These implementations provide the actual logic for data storage and retrieval, enabling the application to interact with various data sources such as databases or external services. Each implementation adheres to the contracts defined in the `repository` interfaces, ensuring consistency and interchangeability within the codebase.
