+++ backend/database/repository/settings.go
@@ -2,7 +2,7 @@ package repository
 
 import (
 	"context"
-	"main/domain/entity"
+	"database/entity"
 )
 
 type SettingsRepo interface {
